[autotest] scheduler: Add trace for HQE lifespan
At the end of an HQE's lifespan, create a trace object recording the
duration of the HQE. By convention, the topmost span in a trace (which
this is) should have an id of 0. We also use by convention a hash of the
HQE's id as the trace id - this allows agents associated with this HQE
to know what the traceId and parentSpanId values are without relying on
global state, or having to be within a
chromite.lib.cloud_trace.SpanStack contextmanager.

BUG=chromium:729199
TEST=None

Change-Id: I2ceba1f5bef87b292f01faf85009599284865fe0
Reviewed-on: https://chromium-review.googlesource.com/655798
Commit-Ready: Paul Hobbs <phobbs@google.com>
Tested-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>