cr50_test: check mosys brand independently of rlz
On some versions of ChromeOS the mosys platform brand is independent of
the rlz code stored in vpd. Check the mosys platform brand separately
from the vpd rlz code.

On these devices the mosys brand cannot be changed from the device
value. Board id tests erase the rlz brand code in hopes of clearing the
mosys brand. This is to prevent cr50-result from trying to set the brand
code on reboot. It is ok if we can't clear the mosys brand code on these
devices with the hardcoded brands, because cr50-result has stopped
trying to set the board id when the device is not running in normal
mode. I am keeping the clear vpd rlz code and the mosys platform brand
checks to be compatible with pre M62 releases.

BUG=none
BRANCH=none
TEST=run firmware_Cr50BID

Change-Id: If109a90b1e0e0765f381af2641c14a40b163fe17
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/619566
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>