[autotest] Improve host dir cleanup.
If the shared host dir already exists when ContainerBucket.setup_base is
called, don't always attempt to clean it up - only do so if the
force_cleanup flag was set, or if the existing directory isn't correctly
set up (e.g. if it's not a shared bind-mount).

This makes running unit tests easier, and also makes it so that the
soon-to-be-implemented container pool service doesn't have to deal with
extraneous exceptions.

BUG=chromium:720219
TEST=sudo python lxc_functional_test.py -v
TEST=sudo python container_bucket_unittest.py -v
TEST=sudo python container_unittest.py -v
TEST=sudo python zygote_unittest.py -v

Change-Id: I926a8e6f36c443b6679edbab622c2bb16c8da14d
Reviewed-on: https://chromium-review.googlesource.com/624699
Commit-Ready: Ben Kwa <kenobi@chromium.org>
Tested-by: Ben Kwa <kenobi@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ben Kwa <kenobi@chromium.org>