Let VirtualFilesystemImage restore loop device ownership and permissions.
VirtualFilesystemImage sets up a loop-mounted virtual filesystem image
for the platform_CrosDisksFilesystem tests to verify the D-Bus API
exposed by cros-disks. As the ownership and permissions of the loop
device may be changed during the tests, this CL updates
VirtualFilesystemImage to properly restore the ownership and permissions
of the loop device.

BUG=None
TEST=Run platform_CrosDisksFilesystem tests and manually check that
VirtualFilesystemImage properly restores the ownership and permissions
of the loop device.

Change-Id: I9e1d23ff12e2b449462df6276643b3b0731a6769
Reviewed-on: https://chromium-review.googlesource.com/624664
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>