network_WiFi_RxFrag: whirlwind APs don't support fragmentation
As noted in, e.g., crosreview.com/333465, no ath10k firmware is known to
support configuring the fragmentation threshold correctly. It's extra
confusing that a packet capture run on the whirlwind will *tell* you
that it has the correct fragmentation fields, but the frames get
corrupted on their way out to the air.

We will hopefully get a better fix for this in crbug.com/753177, when we
can upgrade whirlwind images in the lab and autodetect support for this
feature.

BUG=chromium:597473, chromium:753177
TEST=network_WiFi_RxFrag on whirlwind reports TestNA

Change-Id: I0cc5278c7e42ae64fbb941d9e0d5fdaef88d31ef
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/624733
Reviewed-by: Edward Hill <ecgh@chromium.org>