autotest: Improve error message when choking on unmanaged pool
If a DUT has two pool: labels, one managed and one unmanaged,
lab_inventory may choke on the DUT (depending on which pool
status_history blesses). In this case, we got no information whatsoever
about where the failure was.
This CL is a small improvement in the error message so that the next
person doesn't need to spend an hour chasing the error message.

BUG=chromium:757815
TEST=None

Change-Id: Ia84eaa3d3a1dfaa9e7996d2adbfccfcb4e5f1e61
Reviewed-on: https://chromium-review.googlesource.com/627162
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>