[autotest] Refactor TestPlatform in AU test.
autoupdate_EndToEndTest is badly in need of a refactor. This is the
first of a few planned CLs to do that.

The test file contains two classes (300+ lines) that I will be refactoring
in this CL: TestPlatform and ChromiumOSTestPlatform.

These two are used to represent an "updateable os". It is an implementation
left over from a time when the test supported Brillo and ChromeOS.

I am going to remove the TestPlatform interface class and move
ChromiumOSTestPlatform to a separate file/module.

The separate module will represent the CROS DUT in the test.
It will have functions for things the DUT can do: get logs, reboot, update to
payload X etc

The logic for staging payloads currently lives in ChromiumOSTestPlatform.
I will be moving this code to the test class instead. The "DUT"
shouldn't know about staging paylaods.

So the test class will stage payloads, control the flow of the test, and
check the hostlog of events returned from the update at the end.
ChromiumOSTestPlatform will be used to start the update, reboot, login,
get logs.

BUG=chromium:709710
TEST=autoupdate_EndToEndTest

Change-Id: I9a770bdf9b2c38a83d2801c986184b94f1e0ff01
Reviewed-on: https://chromium-review.googlesource.com/627860
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>