Pick webviews based on the 'screen' url param
cfm_facade_native is a simple API wrapper that issues commands
to a webview. The webview picking logic is not working properly
if there are multiple webviews available (e.g. MIMO + main screen)
and any one of them could be returned.

This change adds the screen param to the constructor to be able
to create cfm facades for specific webviews. It also updates the
necessary logic in kiosk_utils.py, points the _cfm_proxy (in
cfm_facade_adapter) to use the main screen and exposes a new mimo
facade in the xmlproxy that will be used in future changes/tests.

This change should not affect existing tests as all tests that use
the cfm_facade_adapator.cfm since commands are now always sent to
the main window.

TEST=local setup
BUG=chromium:756166

Change-Id: Idb8b7f490ed17c09dc6e01f0ff79a4bf17edff6c
Reviewed-on: https://chromium-review.googlesource.com/628540
Commit-Ready: Claes Malmnäs <malmnas@google.com>
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Claes Malmnäs <malmnas@google.com>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Claes Malmnäs <malmnas@google.com>
Reviewed-by: Denis Tosic <dtosic@chromium.org>