platform_Powerwash: make network connection failures more obvious
This test can fail due to network issues (e.g., "Read from socket
failed: Connection reset by peer") but still end up reporting this
result:

TestFail: Powerwash count didn't increase after powerwash cycle.

That's a bit cryptic when reading the top-level builder results. It's
also bad if we just assume that the value was 0 when it was actually
unknown.

This should all be clearer if we just let the SSH command fail, and
explicitly test for the one failure we want to ignore: when the
POWERWASH_COUNT file is missing.

BUG=chromium:758315
TEST=platform_Powerwash

Change-Id: I3427331d32445f79de8a1223b895ba465d3d2044
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/629756
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>