[autotest] Fix bad argument to test_connection_by_device
The method test_connection_by_device attempts to make the emulated
Bluetooth device connect to the DUT. Consequently it needs to be passed
the emulated device object, not the address of the emulated device.

Note that bluetooth_AdapterPairing.mouse.suspend_resume still fails,
even with this patch. The test has a logic error that this fix exposed.

The logic error is under investigation, and is that the DUT does not
disconnect the emulated device when suspending. Since the DUT is
already connected to the emulated device, (an RN-42-EK connected to
Chameleon,) and the emulated device does not accept commands when
connected, the commands fail on Chameleon. (Autotest reports that these
commands have succeeded, which is incorrect.)

BUG=chromium:724669
TEST=Run the specified test, watch it fail with the same reason, though
the logs show that the underlying problem is different:
$ test_that --board ${BOARD} --args "chameleon_host=${CHAMELEON_IP}" \
${DUT_IP} bluetooth_AdapterPairing.mouse.suspend_resume

Change-Id: Idbae10a3606e12ca29642ab86866ce15eb44a7ce
Reviewed-on: https://chromium-review.googlesource.com/634149
Commit-Ready: Alexander Lent <alent@google.com>
Tested-by: Alexander Lent <alent@google.com>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>