[autotest] Add a conversion in result info for backwards compatibility
For tests running with server-side packaging, the result tools are from older
branches, which has a different format. This change checks for the format
difference and convert it to the latest if needed.

The conversion can be removed after R62 is in stable channel.

BUG=chromium:716218
TEST=merge summary from older format to new one.

Change-Id: I618c290e1a0231a9c1843be020b5c814e02f969e
Reviewed-on: https://chromium-review.googlesource.com/634266
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>