autotest: add provision_Cr50Update
The Cr50 device release prep has become more difficult with the addition
of board id. This test will handle updating Cr50 to the correct image
and setting the board id.

In the future it will also be used to update Cr50 to a image from one of
the Cr50 builders.

Once all the support is in, provision_Cr50Update will allow us to easily
switch between board ids to verify images. You could start the test with
cr50-version:prodsigned-0.0.22, then run the suite with
cr50-version:prodsigned-0.0.23/ZZAF:ffffffff:f700/ZZAF:ff80, and then
later go back to verifying TOT with the label
cr50-version:selfsigned-reef-release/R61-9704.0.0. With these options we
can verify signed images and also verify an image before we sign it.

It is important that we verify the cr50 images before we sign them with
the prod keys. The test suite is supposed to be what catches any
security holes, and we should not only be relying on manually verifying
the changes cherry-picked into the branch. Being able to schedule tests
with images from the builder should help us that.

BUG=none
BRANCH=none
TEST=run provision_Cr50Update

Change-Id: If3b5227d3fab1b7bafd5209e7036e0a6f826bf5d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/648630
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>