[autotest] Add Container IDs.
Containers have identifying information that links them to their test
job.  This information is currently encoded in the container name.  With
the move to zygotes, this scheme no longer works because the identifying
information is not yet available when the zygotes are created, and
running containers cannot be renamed or moved.

Introduce a new ContainerId class that is persistable and enables this
information to be set and stored with the container after it has been
started.

BUG=chromium:720219
TEST=sudo python container_unittest.py
TEST=sudo python container_bucket_unittest.py
TEST=sudo python lxc_functional_test.py

Change-Id: I0a1b427f5fe33f44acc7b37679155b7e879e93cc
Reviewed-on: https://chromium-review.googlesource.com/639730
Commit-Ready: Ben Kwa <kenobi@chromium.org>
Tested-by: Ben Kwa <kenobi@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>