faft: Add unibuild support for fw updates
The fw tests assume that bios.bin and ec.bin are located in the root of
the shellball.  With unified builds, this is no longer true.

Updates the tests to first grep setvars.sh for a given model to
determine where the actual bios/ec images are located.

BUG=b:64465056
TEST=test_that --board=coral 100.100.219.109 firmware_UpdateFirmwareDataKeyVersion
BRANCH=None

Change-Id: I59cf430d1a1a59b580763dd31e377500633d4a2c
Reviewed-on: https://chromium-review.googlesource.com/644320
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>