[autotest] Fix container cloning on moblab.
The code for properly setting snapshot settings was in the
ContainerBucket, so container cloning only worked properly everywhere if
it was invoked via the ContainerBucket.create_from_base function.  In
particular, this causes unit tests to fail on moblab.

Fix the code so that Container.clone works everywhere.

Also, disable some newly-added tests that are guaranteed to fail on
moblab.

BUG=chromium:760735
TEST=sudo python base_image_unittest.py
TEST=sudo python container_unittest.py
TEST=sudo python lxc_functional_test.py

Change-Id: I75e3ffa9a3d6ccf3cf2ad1b720a81d4bf814ec89
Reviewed-on: https://chromium-review.googlesource.com/644322
Commit-Ready: Ben Kwa <kenobi@chromium.org>
Tested-by: Ben Kwa <kenobi@chromium.org>
Reviewed-by: Ben Kwa <kenobi@chromium.org>