cheets_GTS: Mark two modules as no-test modules.
The modules are implemeted in a way that they produce "not-executed"
for all tests when they are run on pre-O devices. Regard such results
as passing.

This treatment is not satisfiable and should be removed in the future.
(See the added comment in the code for the detail.)

BUG=b:65145338
TEST=GtsExternalSourcesTestCases

Change-Id: I6cb480db0e2e641be6e299cb3df0bc74bf421907
Reviewed-on: https://chromium-review.googlesource.com/644929
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>