Add WebRTC stress test for video source switching
This test starts 5 high resolution loopback peer connections. Four of
the video elements are styled to be small, thumbnail sized and one is in
its native size. This is intended to simulate a video conference with
one main feed and several smaller feeds.

The test then frequently swaps the sourceObject of the main feed with a
randomly chosen smaller feed.

Extracted much of the Python test support code to a common location to
be shared with the other WebRTC peer connection tests.

The intention is to later add some more verifications than only checking
if Chrome crashes.

BUG=chromium:760983
TEST=Tested against local DUT
Change-Id: Ia9661d0f65ad0c7d5c22b7e29b37995cce01223f
Reviewed-on: https://chromium-review.googlesource.com/647455
Commit-Ready: Kristoffer Erlandsson <kerl@google.com>
Tested-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Denis Tosic <dtosic@chromium.org>