Refactor PeerConnection into a separate file.
This allows us to reuse this class in other tests. The HTML file is now
generated instead of included in the test dir to allow injecting
arbitrary JavaScript files into it.

Created a utility module in common_lib/cros/webrtc_utils.py for
generating the HTML file and working with the shared JavaScript files.

Some of the boilerplate in the run_once method can hopefully be shared
later as well.

TEST=Local DUT
BUG=chromium:758120
pre-cq-configs: bluestreak-pre-cq
Change-Id: I2939fbc03a96007181138c99de88cd00962a17cb
Reviewed-on: https://chromium-review.googlesource.com/647287
Commit-Ready: Kristoffer Erlandsson <kerl@google.com>
Tested-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Denis Tosic <dtosic@chromium.org>