[autotest] Make the performance pools unmanaged.
Something external to Autotest is adding chameleon DUTs to the
'crosperf' and 'performance' pools.  This is causing lab inventory
scripts to fail, because that creates DUTs that are in both
"managed" and "unmanaged" pools, which the tools can't handle.

This removes the two performance pools from the managed pool list
until we can stop the improper relabeling.

BUG=chromium:757815
TEST=None

Change-Id: I1fdcef327dad0dd15ebd546271e9a8457eb7d221
Reviewed-on: https://chromium-review.googlesource.com/650534
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>