[autotest] Re-enable result reporting logic.
This is a long-winded revert of CL:598667

After that CL landed, CL:602868 privatized an object was earlier passed
into Suite.wait
While the intention was good (not needing to pass in a reporter whenever
one wants to call Suite.wait), the approach made it impossible to
unittest that results were actually reported (and so the revert, which
would have added a check back no longer worked).
This CL achieves the same effect by dependency injection -- Suite
creates and used the default _ResultReporter but allows unittests to
pass in a fake.

TEST=unittests pass.

BUG=chromium:751762
BUG=chromium:751428
BUG=chromium:745894

Change-Id: Id5b03209162201c7be1a7bffca7d164dd487fbfa
Reviewed-on: https://chromium-review.googlesource.com/655623
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>