autotest: Report results from original if rescheduling a test failed.
A suite is sometimes expected to report results when a test fails. But,
if the test is being retried, we delay reporting results, to wait for
results from the retry. Sometimes rescheduling the test can fail, and we
end up not reporting results. Not good. Fixit.

BUG=chromium:751428
TEST=Prayers.

Change-Id: Ibcc1427aac1570cf60b78a5c0b81c2292bae6059
Reviewed-on: https://chromium-review.googlesource.com/655773
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>