p2p_ShareFiles: Fix file write race condition.
Somewhere at the end of this test, a 10 byte files is created and assumed the
p2p-server shares the file (with correct size) back to the simulator. But on
some boards this may flake when the p2p-server shares the file info before its
content is written to disk. This causes the test fail because the expected file
on the DUT was not found.

This patch wraps the call to get_peer_files() in another function to check for
the correctness of the file name and size before complaining. However, if the
file is never shared (due to some unknown problem), the simulator does not raise
an exception on timeout. This additional problem is outside the scope of this
bug.

BUG=chromium:762579
TEST=test_that --board=nyan_kitty chromeos4-row13-rack1-host3.cros p2p_ShareFiles

Change-Id: I3ea8bb5ba3f19f0ba5bb863ee4d67b7bf3fbc939
Reviewed-on: https://chromium-review.googlesource.com/660783
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>