faft: Introduce the ChromeBaseEC class for interacting with the base EC
Abstract the interaction with the base EC in this new class. It is
instantized at the begining of the FirmwareTest object, i.e.
self.base_ec, if the base is connected.

It now only contains the key-press related methods. More methods will
be added later.

BUG=b:66211919
TEST=None, waiting a coming change to add tests to this method.

Change-Id: I62c39aee64e60917fff0f94323275b3742374284
Reviewed-on: https://chromium-review.googlesource.com/676230
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>