Remove staging of control files for au suite.
We no longer need to stage control files as part of the au suite. The
npo test case that this test suite controls is covered as part of the
N2N tests in paygen_au* suites.

This change will remove a lot of manual work required by the test team
to run powerwash tests manually when the sanity stage on the builder
fails and the control file is not generated.

BUG=chromium:758307
BUG=chromium:764038
BUG=chromium:709663
TEST=None

Change-Id: I88a9bf7266c5615deb221d4adb50e5fbf4977d25
Reviewed-on: https://chromium-review.googlesource.com/665366
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>