[autotest] mysql_stats: Fix gauge/counters
Some metrics which were Counters should actually be gauages (row lock time avg
and row lock time current waits are not events, they are gauages). Also, added a
row_lock_waits counter, which is a counter of the total number of queries that
had to wait on a row lock.

BUG=None
TEST=Tested locally

Change-Id: I9fe1f3ffaf9c36e426dfb2b311394158b458436a
Reviewed-on: https://chromium-review.googlesource.com/666228
Commit-Ready: Paul Hobbs <phobbs@google.com>
Tested-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Shuqian Zhao <shuqianz@chromium.org>