network_WiFi_RateControl: only run in conductive cells
The test already tries to detect whether it's running in a conductive
setup (and either exits with TestNA, or logs a warning message,
depending on how certain it is). But we can do better, since the lab
maintains a label for conductive setups. This will prevent scheduling
there at all.

BUG=none
TEST=network_WiFi_RateControl in lab

Change-Id: Ic7b904abcd4eff7127fa41e46683704826ccdd7e
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/669763
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>