rpmserver: Accept logging directory as argument in scripts.
We currently hard code the absolute path of the log file generated by
these scripts in rpm_config.ini
This is not maintainable in the lab, because there is no sane way to
override rpm_config.ini on the servers.

Instead, accept the logging directory via commandline.

BUG=chromium:762233
TEST=(1) unittests
     (2) Run rpmserver daemons via upstart.

Change-Id: I5485ae1a25536783b8f5a84d205f0b74bd5de4bc
Reviewed-on: https://chromium-review.googlesource.com/669946
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>