TCPC: add FAFT test to check FW version
this adds a "CompareChipFwToShellBall" test to FAFT.  it probulates
the DUT and cross-checks the found TCPC firmware with the firmware
blobs in the system image shellball.

if the system image shellball does not contain a relevant TCPC
firmware blob, the test passes for that chip.  the test only fails if
there is a "newer" firmware blob available.

Signed-off-by: Caveh Jalali <caveh@google.com>

TEST=ran FAFT manually on an electro with:
	test_that --autotest_dir /home/caveh/trunk/src/third_party/autotest/files/
	--board=reef 169.254.0.20 'f:.*firmware_CompareChipFwToShellBall/control'
	--args bios=/build/reef/firmware/image.serial.bin

BUG=b:35585700

Change-Id: Iede68f4bdd832d208614ad7e5ec07c81608fb065
Reviewed-on: https://chromium-review.googlesource.com/670103
Commit-Ready: Caveh Jalali <caveh@google.com>
Tested-by: Caveh Jalali <caveh@google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>