[autotest] Save metrics for cros_au triggered from autoupdate_EndToEndTest.
autoupdate_EndToEndTest uses dev_server.auto_update() to trigger cros_au
and start an update. When it does this, the devserver tries to parse a
build_name in the format samus-release/RXX-0000.0.0.

But the test is run with two google storage uris from the
gs://chromeos-releases bucket that don't have a build_name in this
format. So it outputs a message that metrics cannot be parsed and it
will continue anyway.

This test runs a lot so we should save metrics from it. The metrics want
the board, release type, and milestone. The only thing I couldn't get
from the URI was the milestone so I'm getting this from the
instructions.json file stored with the bucket.

BUG=chromium:709710
TEST=autoupdate_EndToEndTest and provision_AutoUpdate pass

Change-Id: I99f00d3ff0cf20487e9fa8ae786674c65ddd9409
Reviewed-on: https://chromium-review.googlesource.com/671218
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>