Explicitly query for the meet telemetry API
By default, we are returning the meet telemetry API if the
hangouts API is not available. This might be prone to errors.
If we fail to load a page properly (e.g. due to a join
error) we would return the meet API although it is acutally
not available.

BUG=chromium:764799
TEST=none
pre-cq-configs: bluestreak-pre-cq

Change-Id: I51009332200ce3bff05d27ea8845707fd623f8cf
Reviewed-on: https://chromium-review.googlesource.com/672343
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>