Takes a screenshot in .png format when test fails
Takes a screenshot if the enterprise_CFM_Sanity test fails
and stores it in the debug dir. Also exposes the necessary APIs
on the native facade & facade adapter. This is to improve
debugging when a test fails. We will later on move this logic
to a common base class for cfm tests.

Logs a warning if screenshotting failed, but does not break the
test.

pre-cq-configs: bluestreak-pre-cq
BUG=chromium:767005
TEST=with local setup & lab device
Change-Id: I203e5852673812a49b9e5e3ee099cb2859503d23
Reviewed-on: https://chromium-review.googlesource.com/672865
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Claes Malmnäs <malmnas@google.com>