Ensure the status check always returns something.
This should fix tests failing with Unhandled EvaluateException:
UncaughtReferenceError, although it is hard to verify since it is not
consistenly reproducable.

We seem to have a race condition where the startCall (which initializes
the testRunner) has not fully completed before the first
testRunner.getStatus() call. This leads to testRunner being undefined
when the status check is made which leads to test failure. See for
example
https://wmatrix.googleplex.com/testrun/unfiltered?test_ids=557813387.

BUG=chromium:766769
TEST=Affected tests against local DUT
pre-cq-configs: bluestreak-pre-cq

Change-Id: I2b321bbdfe204b1f3f8a7e815e2d6e0475af6f4c
Reviewed-on: https://chromium-review.googlesource.com/673445
Commit-Ready: Kristoffer Erlandsson <kerl@google.com>
Tested-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Claes Malmnäs <malmnas@google.com>