[autotest] Improve error reporting for a difficult to debug failure.
Sometimes the test thinks the update failed when it actually succeeded.
This can happen if we lost ssh connection, or the hostname is not
reachable anymore, or if applying the stateful payload failed but rootfs
succeeded. When this happens we retry the update.

But when we retry we are actually starting from a different image that
the test expects (target instead of source).

When this happens the error message wasn't clear enough.

BUG=None
TEST=None

Change-Id: I1b942e3572be65ba8a50171ebec70a902a734441
Reviewed-on: https://chromium-review.googlesource.com/674085
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>