faft_base_ec: Add a test to verify the keyboard work on the base
This is a test similar to firmware_ECKeyboard but talking to BaseEC.
It sends commands to servo and the base EC to emulate key-press. It
then checks if DUT can correctly receive the keyboard events.

The base should be connected to the servo v4 board through an extra
micro-servo.

BUG=b:36075961
TEST=Ran the test on Soraka/Staff locally.

Change-Id: If8e0148d9e3e034dfa1cf07f017b6b4f97214433
Reviewed-on: https://chromium-review.googlesource.com/677885
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>