Fix bugs in sis-updater autotest
Autotest for sis-updater will consistently fail because of some changes
made in the sis-updater logs. Correctely it in this CL.
Also, the autotest used to do a power cycle of Mimo in order to triger
sis-updater. Actually it's not needed because previous updating process
will reset sis and triger it. An extra power cycle has potential of
causing trouble.

BUG=chromium:767642
TEST=test run the autotest for several times and make sure it runs
successfully.

Change-Id: Ied130b637bcd1c0e48b2e97b2989c26d5573d893
Reviewed-on: https://chromium-review.googlesource.com/678820
Commit-Ready: Zhongze Hu <frankhu@google.com>
Tested-by: Zhongze Hu <frankhu@google.com>
Reviewed-by: Zhongze Hu <frankhu@google.com>