autotest: Add retry if devserver call fails due to down service.
It's possible that when devserver service is restarting, it temporarily
doesn't work for a very short time window, but may cause HWTest errors.
This CL adds retry for this case. If after a given time period (60s) the
service is still down, we will declare that this devserver is not serving
and won't retry.

BUG=chromium:766342
TEST=Ran unittest.

Change-Id: I66ed4f138ee89909797b889458ddb490dd5aa304
Reviewed-on: https://chromium-review.googlesource.com/685260
Commit-Ready: Xixuan Wu <xixuan@chromium.org>
Tested-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>