autotest: add cr50_stress_flaky
There are some issues with cr50 tests that require an update/rollback.
We need to workout these issues before we can add the tests to the main
suite. If a test is in the main suite and leaves cr50 running the wrong
version then all following tests in the suite will be invalid.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia74bc69e1ae42f877b72945eba8e2fc701e4c2d1
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/685357
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>