chrome_cr50: change version regexp
Change the version regular expression, so it will match the entire line.
We expect that \d+ will match all of the digits that are after '.', but
we dont necessarily get all of the characters on a line at once. We may
only get the first digit from servo. This is good enough to match, the
expression, but not what we actually want. We sometimes got '0.0.2'
instead of '0.0.27'. This changes the expression to require it to match
a whitespace character at the end of the version. This way we know we
got the entire version string.

Now there are still the occational console command timeouts, but we no
longer get a cutoff version.

BUG=b:66211730
BRANCH=none
TEST=read cr50.get_version() a bunch. Count the number of times it drops
the last character of the version. Verify it is now 0.

Change-Id: Ie9fdfb104df946f1b65dc7c0c18e320476b28a6a
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/687968
Reviewed-by: Wai-Hong Tam <waihong@google.com>