firmware_Cr50BID: dont 'reset state' unless it is necessary
At the start of each test case we currently update to the requested
image and erase the chip bid. Sometimes this is unnecessary, because the
device is already running the correct cr50 image and has an erased chip
board id. This change skips reset_state, if the state is already 'good'.
Skipping an update/rollback saves a lot of time, so this will reduce the
time it takes to run firmware_Cr50BID

BUG=none
BRANCH=none
TEST=none

Change-Id: I0ce602b94a68468a64c64ad84f315579238b49f7
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/693283
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>