firmware_Cr50BID: fix flag tests
Reverse the flag string before trying to get the '1' or '0' index. There
was a mismatch in the endianness of how were getting the first '0' or
'1' from the flags and how were using that index to modify the flags.
This change reverses the flag string before we get the index, so the two
match.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1ebf91ce9e1cb0df6c632ee75ade39253fdb588b
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/693301
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>