firmware_Cr50BID: add option to run subset of test cases
firmware_Cr50BID takes a long time to run, and fails randomly. Add
test_subset, so we can give a list of test cases to run.

If we run firmware_Cr50BID and a test failed because a cr50 console
command timed out, we can rerun that case using the index provided by
the error message.

This is just temporary until we figure out the root cause of the test
issues.

BUG=none
BRANCH=none
TEST=run firmware_Cr50BID --args="test_subset=1,5,6,7" and verify 2, 3,
4, and 8 are skipped.

Change-Id: I7254c5fadc3a99faa1a350a657fd267cef99c35d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/692677
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>