netperf: limit extent of retries, timeout delays
When run with 'ignore_failures' (e.g., with AttenuatedPerf), we forgive
all failure and just wait until the end to collect a None result. This
is pointless and time-wasting, especially on attenuated tests, where we
ramp up attenuation until the client gets abysmal performance (and
likely disconnects from the AP eventually).

We should not keep trying to get 10 samples (MEASUREMENT_MAX_SAMPLES)
when we see multiple None results (i.e., errors). Limit this to 2
failures, and abort on the 3rd.

Also, CL:217970 increased the netperf timeout from a 10 second margin
(on top of test time) to 120 seconds, with no explanation. This seems
excessively high for a 10-second test. And note that we do hit this
timeout sometimes, when running AttenuatedPerf. So let's decrease that
to 60 seconds, which should still be conservative, but should save us
several times that much on some tests, since netperf sessions often
include multiple retries.

BUG=none
TEST=network_WiFi_AttenuatedPerf

Change-Id: I7712fbcb39176c28b037231bb6c4708579672ec4
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/696671
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>