wifi_test_context_manager: log signal level when connected
Test lab issues can show up as poorly-connected networks, with very low
signal strength. Sometimes, we may even be associated only briefly, and
the DUT immediately disconnects. Logging the signal level is a simple
addition to our understanding of failures; it can show if we have a
marginal connection, and it might even help make a quick-disconnect more
obvious (it will print 'None').

BUG=none
TEST=network_WiFi_SimpleConnect.XXX on flaky wifi cells

Change-Id: I3a912fb759b9e59abd5ca18d3dc0ede5fd638067
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/706116
Reviewed-by: Edward Hill <ecgh@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>