[Autotest] Add suite:policy to all policy tests which run in the lab.
To make it easier to check changes for enterprise_policy_base,
put all policy tests into a single suite.  This suite will not
run in the lab, it is just for convenience.  This suite does not
include the generic control files not called by our lab.

Whitelist this attribute as valid.

TEST=ran this new suite
BUG=None

Change-Id: I7fdbaaf8154ccacf9055f9f2d69737a8ad9e8f7c
Reviewed-on: https://chromium-review.googlesource.com/713749
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Krishna Gavini <krishnargv@chromium.org>