cryptohome: check user data visibility after logout
On logout, decrypted user data cached by the file system should
be cleared and become inaccessible. This test checks if a user-
mount file is accessible after the user logs out.

BUG=chromium:769740
TEST="test_that <DUT_IP> platform_CryptohomeDataLeak" passes
when /proc/sys/vm/drop_caches is written to during key
invalidation and fails when it is not written to.

Change-Id: Ic8af4ca85e7ec444e86f882e07cc61ab9479bdcc
Signed-off-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/719576
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>