[autotest] [cfm] Catch all exceptions when taking a screenshot
Before we were catching only StandardError which was not broad enough
causing uncaught exceptions. The specific motivation for this change is
that the screenshot might not be available on the client at all causing
a 'file not found' error while performing scp. Why this fails is still
unclear though.

TEST=local setup
BUG=chromium:772826

Change-Id: Id1834b0d47e1a3c5b1cc110af62c763be228acce
Reviewed-on: https://chromium-review.googlesource.com/720799
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>