random_mac_address: exclude corrupt frames from analysis
A bad FCS might include a bad source HW address, which we use for
determining randomization/non-randomization. Since this might be an
artifact of a bad capture (not a bad DUT behavior), let's exclude these
frames. We should still get enough probe request frames even when
excluding these entries.

This helps particularly with network_WiFi_DisableRandomMACAddress, which
was otherwise getting false negative results where the source address
was getting corrupted in our capture, and so it appeared to be an
unintended randomization.

BUG=none
TEST=network_WiFi_DisableRandomMACAddress, network_WiFi_RandomMACAddress

Change-Id: I144eebe1170340b189cbc09928a3e4898d6824be
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/721574
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>